src/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
Authlib==1.2.1
bertopic==0.15.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
Cython==0.29.36
distro==1.8.0
exceptiongroup==1.1.3
fastapi==0.104.1
fonttools==4.44.0
h11==0.14.0
hdbscan==0.8.33
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.41.1
matplotlib==3.8.1
numba==0.58.1
numpy==1.26.1
openai==1.2.3
packaging==23.2
pandas==2.1.3
patsy==0.5.3
piecewise-regression==1.3.0
Pillow==10.1.0
plotly==5.18.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pymannkendall==1.4.3
pynndescent==0.5.10
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
statsmodels==0.14.0
tbb==2021.10.0
tenacity==8.2.3
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
umap-learn==0.5.4
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.22.0
watchfiles==0.21.0
weaviate-client==3.25.3
websockets==12.0
